Add CancelRun method to Command
This method will nil out the Run and RunE of a command. This can be called from
PreRun-style functions to prevent the command from running.

Example: A user can check if the environment has been initialized in a PreRun
method and, if not, prevent the command from running.

Note that this does not nil out any PostRun-style commands, as these are
typically used for clean-up and tear-down operations that likely still need to
occur.

Merge https://github.com/spf13/cobra/pull/657

Signed-off-by: Kevin C. Wells <kevin.c.wells@intel.com>