Clarify documentation for completion
It was pointed out on Slack that there is no explanation in the docs
of what the `toComplete` parameter represents for `ValidArgsFunction`.

Also, I felt that using the word `annotation` when talking about
descriptions of completions could be confusing, since Cobra does
have a concept of annotations for completions, which represent
something else.

This PR tries to improve the docs on both points.

Signed-off-by: Marc Khouzam <marc.khouzam@montreal.ca>